move disable button logic to seriesepisode component
